requirements.txt
gunicorn==20.1.0
backcall==0.2.0
Brotli==1.0.9
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
cloudpickle==2.1.0
cycler==0.11.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
executing==0.8.3
Flask==2.1.2
fonttools==4.33.3
idna==3.3
imbalanced-learn==0.9.1
imblearn==0.0
importlib-metadata==4.11.4
interpret-core==0.2.7
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
kiwisolver==1.4.2
llvmlite==0.38.1
MarkupSafe==2.1.1
matplotlib==3.5.2
numba==0.55.1
numpy==1.22.4
packaging==21.3
pandas==1.4.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
pox==0.3.1
ppft==1.7.6.5
prompt-toolkit==3.0.29
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.12.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
pyzmq==23.0.0
requests==2.27.1
scikit-learn==1.1.1
scipy==1.8.1
shap==0.40.0
six==1.16.0
slicer==0.0.7
tenacity==8.0.1
threadpoolctl==3.1.0
torch==1.11.0
tornado==6.1
tqdm==4.64.0
traitlets==5.2.1.post0
treeinterpreter==0.2.3
typing_extensions==4.2.0
urllib3==1.26.9
wcwidth==0.2.5
Werkzeug==2.1.2
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.4.0
